Finally found bug with viewing full sized starred items
Basically the MediaRow constructor in the IndexOf call was defaulting to starred = false and thus it couldn't find that object in the List and thus index was -1. No matter what it only ever displayed the right arrow (next) btn and what would happen is the index would be set to -1 and when added to 1 would get 0, thus it would always take you to start of list. This took way too long to debug.